requirements.txt
absl-py==2.1.0
annotated-types==0.5.0  
asttokens==2.4.1
astunparse==1.6.3
atomicwrites==1.4.1
attrs==23.2.0
backcall==0.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
clean-fid==0.1.35
colorama==0.4.6
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
docstring_parser==0.16
eval_type_backport==0.2.0
executing==2.0.1
filelock==3.15.4
flake8==3.8.1
flake8-import-order==0.18.1
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.6.1
fvcore==0.1.5.post20221221
gast==0.4.0
google-auth==2.29.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.20.3
grpcio==1.63.0
gviz-api==1.10.0
h5py==3.11.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
iniconfig==2.0.0
intel-openmp==2021.4.0
iopath==0.1.9 
ipykernel==6.21.3
ipython==7.34.0  
isort==4.3.21
jedi==0.18.2 
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.1
jupyter_core==5.7.2
keras==2.13.1
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mccabe==0.6.1
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.1
numpy==1.24.3
nvidia-ml-py==12.555.43
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==24.0
pandas==2.0.3
parso==0.8.3  
pickleshare==0.7.5
pillow==10.3.0
platformdirs==4.2.1
pluggy==1.5.0
portalocker==2.10.1
progress==1.6
prompt-toolkit==3.0.38 
protobuf==4.25.4
psutil==5.9.8
ptflops==0.7.2.2
pure-eval==0.2.2
py==1.11.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.6.0
pydantic==2.7.0  
pydantic_core==2.20.1
pyflakes==2.2.0
Pygments==2.18.0
pynvml==11.5.3
pyparsing==3.1.2
pytest==7.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
shtab==1.7.1
six==1.16.0
spikingjelly==0.0.0.0.12
stack-data==0.6.3
sympy==1.13.1
syops==0.0.5
tabulate==0.9.0
tbb==2021.13.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorboard-plugin-profile==2.15.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.5.0
tomli==2.0.1
torch-tb-profiler==0.4.3
torchprofile==0.0.4
torchsummary==1.5.1
torchviz==0.0.2
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
typing==3.7.4.3
typing_extensions==4.5.0 
tyro==0.8.5
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.2
wrapt==1.16.0
yacs==0.1.8
zeus-ml==0.8.2
zipp==3.18.1
